requirements.txt
fastapi 
uvicorn 
azure-ai-formrecognizer 
python-multipart
streamlit
python-docx
chromadb
langchain
PyPDF2
ollama
azure-ai-documentintelligence==1.0.0b4
python-dotenv